Resolved conflicts
