+++ package/zustand_generator/src/component/Editor.tsx
@@ -1,154 +1,27 @@
-import React, {useState} from 'react';
+import React, { useState } from 'react';
 import CodeEditor from '@uiw/react-textarea-code-editor';
 import NumberList from './Numbers';
-// import { generateTests } from './component-utils';
+import { Header } from './Header';
 
 const editorStyle: React.CSSProperties = {
   display: 'flex',
   flexDirection: 'column',
-  height: '100vh',
+  height: '100%',
   overflow: 'hidden',
+  borderLeft: '1px solid rgba(243,246,248,.1)',
   backgroundColor: '#1C1C1C',
 };
-
-const toolBar: React.CSSProperties = {
-  display: 'flex',
-  height: '56px',
-  width: '667px',
-  padding: '8px 16px',
-  alignItems: 'center',
-  fontFamily: 'Inter',
-  gap: '8px',
-}
-
 const codePanel: React.CSSProperties = {
   display: 'flex',
-  border: '1px solid blue',
   flexGrow: 1,
-  overflow: 'auto',
-  height: 'calc(100vh - 200px)',
-}
-
-const toolBarLogoBox: React.CSSProperties = {
-  height: '40px',
-  width: '40px',
-  backgroundColor: 'white',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-}
-
-const toolBarLogo: React.CSSProperties = {
-  // insert image
-}
-
-const toolBarTitleContainer: React.CSSProperties = {
-  display: 'flex',
-  flexDirection: 'column',
-  color: '#C6C6C6',
-  lineHeight: '16px',
-  opacity: 0.9,
-  fontFamily: 'Inter',
-  marginLeft: '4px',
-  flexGrow: 1,
-}
-
-const toolBarTitle: React.CSSProperties = {
-  fontSize: '16px',
-  fontWeight: 500,
-  fontFamily: 'Inter !important',
-}
-
-const toolBarDescription: React.CSSProperties = {
-  fontSize: '12px',
-  fontWeight: 400,
-}
-
-const downloadButton: React.CSSProperties = {
-  height: '32px',
-  width: '138px',
-  backgroundColor: 'rgb(243, 246, 248, 0.03)',
-  border: '1px solid rgba(243, 246, 248, 0.05)',
-  borderRadius: '6px',
-  display: 'flex',
-  flexDirection: 'row',
-  alignItems: 'center',
-  padding: '6px 6px 6px 12px',
-  gap: '4px',
-}
-
-const downloadTitleContainer: React.CSSProperties = {
-  display: 'flex',
-  alignItems: 'center',
-  gap: '10px',
-}
-
-const downloadIconBox: React.CSSProperties = {
-  backgroundColor: 'white',
-  height: '20px',
-  width: '20px',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-}
-
-const downloadIcon: React.CSSProperties = {
-  // insert image
-}
-
-const downloadTitleText: React.CSSProperties = {
-  fontSize: '12px',
-  fontWeight: 500,
-  lineHeight: '16px',
-  color: 'rgba(243, 246, 248, 0.7)'
-}
-
-const downloadLine: React.CSSProperties = {
-  width: '0px',
-  height: '20px',
-  border: '1px solid rgba(243, 246, 248, 0.05)',
-}
-
-const downloadArrow: React.CSSProperties = {
-  width: '20px',
-  height: '20px',
-  backgroundColor: 'red',
-} 
-
-const copyButton: React.CSSProperties = {
-  height: '32px',
-  width: '32px',
-  backgroundColor: 'rgb(243, 246, 248, 0.03)',
-  border: '1px solid rgba(243, 246, 248, 0.05)',
-  borderRadius: '6px',
-  display: 'flex',
-  alignItems: 'center',
-  padding: '6px',
-}
-
-const minimizeButton: React.CSSProperties = {
-  height: '24px',
-  width: '24px',
-  backgroundColor: 'white',
-  marginLeft: '10px',
-  padding: 0,
-  cursor: 'pointer',
-}
-
-const minimizeIcon: React.CSSProperties = {
-
-}
-
-const copyIcon: React.CSSProperties = {
-  // add icon here
-}
+  overflowY: 'scroll',
+};
 
 interface Props {
   code: string;
-  setIsHidden: Function,
-  isHidden: boolean,
+  setIsHidden: Function;
+  isHidden: boolean;
 }
-
 const Editorfield = ({ code, isHidden, setIsHidden }: Props): JSX.Element => {
   // const [storeMap] = React.useState<Map<string, string>>(new Map());
   const [, setInnerCode] = useState(code);
@@ -162,52 +35,27 @@ const Editorfield = ({ code, isHidden, setIsHidden }: Props): JSX.Element => {
 
   return (
     <div style={editorStyle}>
-      <div style={toolBar}>
-        <div style={toolBarLogoBox}>
-          <div style={toolBarLogo}/>
-        </div>
-        <div style={toolBarTitleContainer}>
-          <div style={toolBarTitle}>Chromogen Tests</div>
-          <div style={toolBarDescription}>Interact with the app to generate tests</div>
-        </div>
-        <button type="button" style={downloadButton}>
-          <div style={downloadTitleContainer}>
-            <div style={downloadIconBox}>
-              <div style={downloadIcon}/>
-            </div>
-            <div style={downloadTitleText}>Download</div>
-          </div>
-          <div style={downloadLine}></div>
-          <div style={downloadArrow}></div>
-        </button>
-        <button type="button" style={copyButton}>
-          <div style={copyIcon}/>
-        </button>
-        <button type="button" style={minimizeButton} onClick={() => setIsHidden(!isHidden)}>
-          <div style={minimizeIcon}/>
-        </button>
-      </div>
+      <Header isHidden={isHidden} setIsHidden={setIsHidden} />
       <div style={codePanel}>
-      <NumberList number={breakLine + 10} />
-      <CodeEditor
-        data-color-mode="dark"
-        value={code}
-        language="js"
-        placeholder="Please enter JS code."
-        onChange={(evn) => setInnerCode(evn.target.value)}
-        padding={15}
-        style={{
-          maxWidth: 1000,
-          width: '40wv',
-          fontSize: 12,
-          backgroundColor: '#1c1c1c',
-          fontFamily:
-            'ui-monospace, SFMono-Regular, SF Mono, Consolas, Liberation Mono, Menlo,monospace',
-        }}
+        <NumberList number={breakLine + 10} />
+        <CodeEditor
+          data-color-mode="dark"
+          value={code}
+          language="js"
+          placeholder="Please enter JS code."
+          onChange={(evn) => setInnerCode(evn.target.value)}
+          padding={15}
+          style={{
+            maxWidth: 1000,
+            width: '40wv',
+            fontSize: 12,
+            backgroundColor: '#1c1c1c',
+            fontFamily:
+              'ui-monospace, SFMono-Regular, SF Mono, Consolas, Liberation Mono, Menlo,monospace',
+          }}
         />
-        {/* <button onClick={() => console.log(generateTests(storeMap)[0])}>Click me</button> */}
       </div>
-      <div style={toolBar}/>
+      <div style={{ position: 'absolute' }} />
     </div>
   );
 };
